requirements.txt
nltk==3.8.1
pyyaml==6.0
tqdm==4.66.1
python-dateutil==2.8.2 